TODO arreglar newsInterface dbservice

+++ toDoApp/src/app/register/register.component.ts
@@ -16,17 +16,22 @@ import { User } from 'firebase/auth';
 export class RegisterComponent {
   email: string = '';
   password: string = '';
+  confirmPassword: string = '';
   errorMessage: string = '';
 
   constructor(private authService: AuthService, private router: Router) {}
 
   onRegister() {
+    if (this.password !== this.confirmPassword) {
+      this.errorMessage = 'Passwords do not match';
+      return;
+    }
     this.authService.register(this.email, this.password)
       .then((user: User) => {
         // Navigate to the login page or home page after successful registration
         localStorage.setItem('token', user.uid); // Store the token in localStorage
         localStorage.setItem('userName', user.email || 'User'); // Store the user name in localStorage
-        this.router.navigate(['/login']);
+        this.router.navigate(['/home']);
       })
       .catch(error => {
         // Handle registration errors
